Extract image MCP tools
Move the three image analysis MCP tools into `server_tools_image`: extract colors,
generate palette, and analyze product. `server.py` stays as the public façade
and re-exports the moved handlers while importing the split module once for
registration.

Constraint: Preserve MCP tool names and direct imports from `mcp_video.server`
Rejected: Combine image tools with Remotion/AI tools | keep server splits small and easy to verify
Confidence: high
Scope-risk: narrow
Tested: python3 -m pytest tests/test_public_surface.py tests/test_server.py::TestServerInitialization -q --tb=short; python3 -m ruff check mcp_video/server.py mcp_video/server_tools_image.py tests/test_public_surface.py; python3 -m ruff format --check mcp_video/server.py mcp_video/server_tools_image.py; async smoke of mcp.list_tools() returning 83 tools including moved names
Not-tested: full non-slow suite after image tool extraction
